Break down webinar structure into files
