+++ samples/webinar/package.json
@@ -6,9 +6,7 @@
     "build": "npm run build:ts",
     "build:ts": "tsc",
     "build:ts:watch": "tsc  --watch",
-    "start": "run-p start:**",
-    "start:server": "ts-node -r tsconfig-paths/register ./src/index.ts",
-    "start:subscription": "ts-node -r tsconfig-paths/register ./src/subscription.ts",
+    "start": "ts-node -r tsconfig-paths/register ./src/index.ts",
     "test": "npm run test:unit && npm run test:int && npm run test:api && npm run test:e2e",
     "test:unit": "jest unit --passWithNoTests",
     "test:int": "jest int --passWithNoTests",
@@ -40,7 +38,7 @@
   },
   "homepage": "https://github.com/oskardudycz/EventSourcing.NodeJS#readme",
   "dependencies": {
-    "@eventstore/db-client": "file:libs/eventstore-db-client-3.2.0-alpha.1.tgz",
+    "@eventstore/db-client": "3.2.0",
     "convict": "6.2.1",
     "dotenv": "16.0.0",
     "express": "4.17.3",
