+++ samples/webinar/src/shoppingCarts/productItem.ts
@@ -1,0 +1,71 @@
+//////////////////////////////////////
+/// Product Items
+//////////////////////////////////////
+
+import { ShoppingCartErrors } from './shoppingCart';
+
+export interface ProductItem {
+  productId: string;
+  quantity: number;
+}
+
+export const addProductItem = (
+  productItems: ProductItem[],
+  newProductItem: ProductItem
+): ProductItem[] => {
+  const { productId, quantity } = newProductItem;
+
+  const currentProductItem = findProductItem(productItems, productId);
+
+  if (!currentProductItem) return [...productItems, newProductItem];
+
+  const newQuantity = currentProductItem.quantity + quantity;
+  const mergedProductItem = { productId, quantity: newQuantity };
+
+  return productItems.map((pi) =>
+    pi.productId === productId ? mergedProductItem : pi
+  );
+};
+
+export const removeProductItem = (
+  productItems: ProductItem[],
+  newProductItem: ProductItem
+): ProductItem[] => {
+  const { productId, quantity } = newProductItem;
+
+  const currentProductItem = assertProductItemExists(
+    productItems,
+    newProductItem
+  );
+
+  const newQuantity = currentProductItem.quantity - quantity;
+
+  if (newQuantity === 0)
+    return productItems.filter((pi) => pi.productId !== productId);
+
+  const mergedProductItem = { productId, quantity: newQuantity };
+
+  return productItems.map((pi) =>
+    pi.productId === productId ? mergedProductItem : pi
+  );
+};
+
+export const findProductItem = (
+  productItems: ProductItem[],
+  productId: string
+): ProductItem | undefined => {
+  return productItems.find((pi) => pi.productId === productId);
+};
+
+export const assertProductItemExists = (
+  productItems: ProductItem[],
+  { productId, quantity }: ProductItem
+): ProductItem => {
+  const current = findProductItem(productItems, productId);
+
+  if (!current || current.quantity < quantity) {
+    throw ShoppingCartErrors.PRODUCT_ITEM_NOT_FOUND;
+  }
+
+  return current;
+};
